move out string
