+++ rust/blockstore/src/arrow/block/delta/storage.rs
@@ -1,3 +1,4 @@
+use super::string::StringValueStorage;
 use crate::{
     arrow::types::ArrowWriteableKey,
     key::{CompositeKey, KeyWrapper},
@@ -21,7 +22,7 @@ use std::{
 };
 
 #[derive(Clone)]
-pub enum BlockStorage {
+pub(in crate::arrow) enum BlockStorage {
     String(StringValueStorage),
     Int32Array(Int32ArrayStorage),
     UInt32(UInt32Storage),
@@ -49,7 +50,7 @@ pub enum BlockKeyArrowBuilder {
 }
 
 impl BlockKeyArrowBuilder {
-    fn add_key(&mut self, key: CompositeKey) {
+    pub(super) fn add_key(&mut self, key: CompositeKey) {
         match key.key {
             KeyWrapper::String(value) => {
                 let builder = match self {
@@ -145,130 +146,6 @@ impl BlockKeyArrowBuilder {
                 )
             }
         }
-    }
-}
-
-#[derive(Clone)]
-pub struct StringValueStorage {
-    pub storage: Arc<RwLock<Option<BTreeMap<CompositeKey, String>>>>,
-}
-
-impl StringValueStorage {
-    pub fn new() -> Self {
-        Self {
-            storage: Arc::new(RwLock::new(Some(BTreeMap::new()))),
-        }
-    }
-
-    fn get_prefix_size(&self, start: usize, end: usize) -> usize {
-        let storage = self.storage.read();
-        match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => {
-                let key_stream = storage
-                    .iter()
-                    .skip(start)
-                    .take(end - start)
-                    .map(|(key, _)| key);
-                calculate_prefix_size(key_stream)
-            }
-        }
-    }
-
-    fn get_key_size(&self, start: usize, end: usize) -> usize {
-        let storage = self.storage.read();
-        match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => {
-                let key_stream = storage
-                    .iter()
-                    .skip(start)
-                    .take(end - start)
-                    .map(|(key, _)| key);
-                calculate_key_size(key_stream)
-            }
-        }
-    }
-
-    pub fn get_value_size(&self, start: usize, end: usize) -> usize {
-        let storage = self.storage.read();
-        match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => {
-                let value_stream = storage
-                    .iter()
-                    .skip(start)
-                    .take(end - start)
-                    .map(|(_, value)| value);
-                value_stream.fold(0, |acc, value| acc + value.len())
-            }
-        }
-    }
-
-    pub fn len(&self) -> usize {
-        let storage = self.storage.read();
-        match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => storage.len(),
-        }
-    }
-
-    fn build_keys(&self, builder: BlockKeyArrowBuilder) -> BlockKeyArrowBuilder {
-        let storage = self.storage.read();
-        match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => {
-                let mut builder = builder;
-                for (key, _) in storage.iter() {
-                    builder.add_key(key.clone());
-                }
-                builder
-            }
-        }
-    }
-
-    fn split(&self, prefix: &str, key: KeyWrapper) -> StringValueStorage {
-        let mut storage = self.storage.write();
-        match storage.as_mut() {
-            None => unreachable!("Invariant violation. A StringValueBuilder should have storage."),
-            Some(storage) => {
-                let split = storage.split_off(&CompositeKey {
-                    prefix: prefix.to_string(),
-                    key,
-                });
-                StringValueStorage {
-                    storage: Arc::new(RwLock::new(Some(split))),
-                }
-            }
-        }
-    }
-
-    fn to_arrow(&self) -> (Field, ArrayRef) {
-        let item_capacity = self.len();
-        let mut value_builder;
-        if item_capacity == 0 {
-            value_builder = StringBuilder::new();
-        } else {
-            value_builder =
-                StringBuilder::with_capacity(item_capacity, self.get_value_size(0, self.len()));
-        }
-
-        let storage = self.storage.read();
-        let storage = match storage.as_ref() {
-            None => unreachable!("Invariant violation. A StringDeltaBuilder should have storage."),
-            Some(storage) => storage,
-        };
-
-        for (_, value) in storage.iter() {
-            value_builder.append_value(value);
-        }
-
-        let value_field = Field::new("value", arrow::datatypes::DataType::Utf8, false);
-        let value_arr = value_builder.finish();
-        (
-            value_field,
-            (&value_arr as &dyn Array).slice(0, value_arr.len()),
-        )
     }
 }
 
@@ -948,10 +825,14 @@ impl BlockStorage {
     }
 }
 
-fn calculate_prefix_size<'a>(composite_key_iter: impl Iterator<Item = &'a CompositeKey>) -> usize {
+pub(super) fn calculate_prefix_size<'a>(
+    composite_key_iter: impl Iterator<Item = &'a CompositeKey>,
+) -> usize {
     composite_key_iter.fold(0, |acc, key| acc + key.prefix.len())
 }
 
-fn calculate_key_size<'a>(composite_key_iter: impl Iterator<Item = &'a CompositeKey>) -> usize {
+pub(super) fn calculate_key_size<'a>(
+    composite_key_iter: impl Iterator<Item = &'a CompositeKey>,
+) -> usize {
     composite_key_iter.fold(0, |acc, key| acc + key.key.get_size())
 }
